add component md2html to resolve duplicated code of index.md.jsx
