+++ src/components/md2html/index.scss
@@ -2,7 +2,7 @@
 @import '../../reset.scss';
 @import '../../markdown.scss';
 
-.community-page {
+.md2html {
   .content-section {
     max-width: $contentWidth;
     margin: 0 auto;
