--- src/pages/documentation/index.md.scss
@@ -1,36 +1,0 @@
-@import '../../variables.scss';
-@import '../../reset.scss';
-@import '../../markdown.scss';
-
-.documentation-page {
-  .content-section {
-    max-width: $contentWidth;
-    margin: 0 auto;
-    box-sizing: border-box;
-    padding: 40px 40px 60px;
-    position: relative;
-    min-height: 1100px;
-    .doc-content {
-      display: inline-block;
-      vertical-align: top;
-      box-sizing: border-box;
-      padding: 20px 6% 0;
-      width: calc(100% - #{$sideMenuWidth});
-    }
-  }
-  @media screen and (max-width: $mobileWidth) {
-    .content-section {
-      padding-left: 20px;
-      padding-right: 20px;
-      .doc-content {
-        width: 100%;
-      }
-    }
-    .sidemenu {
-      position: absolute;
-      z-index: 100;
-      left: 0;
-      top: 40px;
-    }
-  }
-}
